+++ src/main/java/com/kakaotech/team14backend/inner/post/usecase/FindPersonalPostListUsecase.java
@@ -2,7 +2,7 @@ package com.kakaotech.team14backend.inner.post.usecase;
 
 import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
-import com.kakaotech.team14backend.outer.post.dto.GetPersonalPostListResponseDTO;
+import com.kakaotech.team14backend.post.dto.GetPersonalPostListResponseDTO;
 import com.kakaotech.team14backend.post.application.PostMapper;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
