+++ src/main/java/com/kakaotech/team14backend/inner/post/usecase/FindPostListUsecase.java
@@ -3,8 +3,8 @@ package com.kakaotech.team14backend.inner.post.usecase;
 import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.inner.post.repository.PostLikeRepository;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
-import com.kakaotech.team14backend.outer.post.dto.GetHomePostListResponseDTO;
-import com.kakaotech.team14backend.outer.post.dto.SetAuthenticatedHomePostDTO;
+import com.kakaotech.team14backend.post.dto.GetHomePostListResponseDTO;
+import com.kakaotech.team14backend.post.dto.SetAuthenticatedHomePostDTO;
 import com.kakaotech.team14backend.post.application.PostMapper;
 import com.kakaotech.team14backend.outer.post.service.FindLikeStatusService;
 import java.util.ArrayList;
