+++ src/main/java/com/kakaotech/team14backend/inner/post/usecase/FindPostUsecase.java
@@ -2,8 +2,8 @@ package com.kakaotech.team14backend.inner.post.usecase;
 
 import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
-import com.kakaotech.team14backend.outer.post.dto.GetPostDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetPostResponseDTO;
+import com.kakaotech.team14backend.post.dto.GetPostDTO;
+import com.kakaotech.team14backend.post.dto.GetPostResponseDTO;
 import com.kakaotech.team14backend.post.application.PostMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
