+++ src/main/java/com/kakaotech/team14backend/inner/post/usecase/UpdatePostLikeCountUsecase.java
@@ -2,7 +2,7 @@ package com.kakaotech.team14backend.inner.post.usecase;
 
 import com.kakaotech.team14backend.inner.post.model.PostLikeCount;
 import com.kakaotech.team14backend.inner.post.repository.PostLikeCountRepository;
-import com.kakaotech.team14backend.outer.post.dto.GetPostLikeCountDTO;
+import com.kakaotech.team14backend.post.dto.GetPostLikeCountDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
