--- src/main/java/com/kakaotech/team14backend/outer/post/dto/GetHomePostInterface.java
@@ -1,4 +1,0 @@
-package com.kakaotech.team14backend.outer.post.dto;
-
-public interface GetHomePostInterface {
-}
