+++ src/main/java/com/kakaotech/team14backend/post/dto/GetHomePostInterface.java
@@ -1,0 +1,4 @@
+package com.kakaotech.team14backend.post.dto;
+
+public interface GetHomePostInterface {
+}
