+++ src/main/java/com/kakaotech/team14backend/post/dto/GetPersonalPostResponseDTO.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.outer.post.dto;
+package com.kakaotech.team14backend.post.dto;
 
 public record GetPersonalPostResponseDTO(Long postId, String imageUri, String nickname,
                                          String createdAt, Long viewCount, Long likeCount) {
