+++ src/main/java/com/kakaotech/team14backend/post/dto/GetPostDTO.java
@@ -1,4 +1,4 @@
-package com.kakaotech.team14backend.outer.post.dto;
+package com.kakaotech.team14backend.post.dto;
 
 public record GetPostDTO(Long postId, Long memberId) {
 }
