+++ src/main/java/com/kakaotech/team14backend/post/presentation/PostController.java
@@ -11,18 +11,18 @@ import com.kakaotech.team14backend.post.exception.MaxLevelSizeException;
 import com.kakaotech.team14backend.exception.SizeParameterException;
 import com.kakaotech.team14backend.exception.UserNotAuthenticatedException;
 import com.kakaotech.team14backend.inner.member.model.Member;
-import com.kakaotech.team14backend.outer.post.dto.GetHomePostListResponseDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetMyPostResponseDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetPersonalPostListResponseDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetPopularPostListRequestDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetPopularPostListResponseDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetPopularPostResponseDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetPostDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetPostResponseDTO;
-import com.kakaotech.team14backend.outer.post.dto.SetPostLikeDTO;
-import com.kakaotech.team14backend.outer.post.dto.SetPostLikeResponseDTO;
-import com.kakaotech.team14backend.outer.post.dto.UploadPostDTO;
-import com.kakaotech.team14backend.outer.post.dto.UploadPostRequestDTO;
+import com.kakaotech.team14backend.post.dto.GetHomePostListResponseDTO;
+import com.kakaotech.team14backend.post.dto.GetMyPostResponseDTO;
+import com.kakaotech.team14backend.post.dto.GetPersonalPostListResponseDTO;
+import com.kakaotech.team14backend.post.dto.GetPopularPostListRequestDTO;
+import com.kakaotech.team14backend.post.dto.GetPopularPostListResponseDTO;
+import com.kakaotech.team14backend.post.dto.GetPopularPostResponseDTO;
+import com.kakaotech.team14backend.post.dto.GetPostDTO;
+import com.kakaotech.team14backend.post.dto.GetPostResponseDTO;
+import com.kakaotech.team14backend.post.dto.SetPostLikeDTO;
+import com.kakaotech.team14backend.post.dto.SetPostLikeResponseDTO;
+import com.kakaotech.team14backend.post.dto.UploadPostDTO;
+import com.kakaotech.team14backend.post.dto.UploadPostRequestDTO;
 import com.kakaotech.team14backend.post.application.PostService;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
