+++ src/test/java/com/kakaotech/team14backend/inner/post/usecase/UpdatePostLikeCountUsecaseTest.java
@@ -5,9 +5,9 @@ import com.kakaotech.team14backend.inner.member.repository.MemberRepository;
 import com.kakaotech.team14backend.inner.post.model.PostLikeCount;
 import com.kakaotech.team14backend.inner.post.repository.PostLikeCountRepository;
 import com.kakaotech.team14backend.inner.post.repository.PostLikeRepository;
-import com.kakaotech.team14backend.outer.post.dto.GetPostLikeCountDTO;
-import com.kakaotech.team14backend.outer.post.dto.SetPostLikeDTO;
-import com.kakaotech.team14backend.outer.post.dto.SetPostLikeResponseDTO;
+import com.kakaotech.team14backend.post.dto.GetPostLikeCountDTO;
+import com.kakaotech.team14backend.post.dto.SetPostLikeDTO;
+import com.kakaotech.team14backend.post.dto.SetPostLikeResponseDTO;
 import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.post.infrastructure.PostRepository;
 import org.junit.jupiter.api.Test;
